+++ src/main/java/br/edu/infnet/erik/model/repository/StakeholderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StakeholderRepository extends CrudRepository<Stakeholder, Integer> {
+public interface StakeholderRepository extends CrudRepository<Stakeholder, Long> {
 
     Stakeholder findByCadastroPessoa(String cadastroPessoa);
 }
